Change role and user id's to be optional.  When creating these objects the id's should be null, not 0.

+++ src/org/hyperic/hq/hqapi1/test/RoleGetRole_test.java
@@ -18,7 +18,7 @@ public class RoleGetRole_test extends RoleTestBase {
         hqAssertSuccess(response);
 
         Role r = response.getRole();
-        assertEquals(2, r.getId());
+        assertEquals(2, r.getId().intValue());
     }
 
     public void testGetRoleNameInvalid() throws Exception {
